+++ src/Service/RustService.php
@@ -10,6 +10,7 @@ use Castor\Context;
 use Castor\Docker\Service\Behaviour\HasDirectory;
 use Castor\Docker\Service\Behaviour\HasDockerfile;
 use Castor\Docker\Service\Behaviour\HasHttpRouting;
+use Castor\Docker\Service\Behaviour\HasLinks;
 use Castor\Docker\Service\Behaviour\HasSharedHomeDirectory;
 use Castor\Docker\Service\Behaviour\HasVersion;
 use Castor\Docker\Service\Builder\ComposeBuilder;
@@ -46,6 +47,7 @@ class RustService implements ServiceInterface
     use HasDirectory;
     use HasDockerfile;
     use HasHttpRouting;
+    use HasLinks;
     use HasSharedHomeDirectory;
     use HasVersion;
 
@@ -234,6 +236,7 @@ class RustService implements ServiceInterface
                 ->environment('CARGO_HOME', '/home/app/.cargo')
         ;
 
+        $this->applyLinks($context, $appService);
         $this->applyBuild($appService, $context);
         $this->applyHttpRouting($appService);